+++ text/message_verification.go
@@ -22,7 +22,7 @@ func NewErrorValidationVerificationFlowExpired(expiredAt time.Time) *Message {
 func NewInfoSelfServiceVerificationSuccessful() *Message {
 	return &Message{
 		ID:   InfoSelfServiceVerificationSuccessful,
-		Type: Info,
+		Type: Success,
 		Text: "You successfully verified your email address.",
 	}
 }
@@ -68,6 +68,15 @@ func NewErrorValidationVerificationCodeInvalidOrAlreadyUsed() *Message {
 		ID:      ErrorValidationVerificationCodeInvalidOrAlreadyUsed,
 		Text:    "The verification code is invalid or has already been used. Please try again.",
 		Type:    Error,
+		Context: context(nil),
+	}
+}
+
+func NewVerificationEmailWithCodeSent() *Message {
+	return &Message{
+		ID:      InfoSelfServiceVerificationEmailWithCodeSent,
+		Type:    Info,
+		Text:    "An email containing a verification code has been sent to the email address you provided.",
 		Context: context(nil),
 	}
 }