+++ text/type.go
@@ -5,12 +5,14 @@ package text
 
 import "time"
 
-// swagger:model uiTextType
-type Type string
+// swagger:enum UITextType
+type UITextType string
 
+// aligned with https://github.com/ory/elements/blob/main/src/theme/message.css.ts
 const (
-	Info  Type = "info"
-	Error Type = "error"
+	Info    UITextType = "info"
+	Error   UITextType = "error"
+	Success UITextType = "success"
 )
 
 var Now = time.Now